refactor: create @europeana/oembed package
For oEmbed client & providers, moved out of plugins dir
