+++ packages/portal/src/components/item/ItemEmbedCode.vue
@@ -27,7 +27,7 @@
 </template>
 
 <script>
-  import { oEmbedForEndpoint } from '@/plugins/oembed';
+  import { oEmbedForEndpoint } from '@europeana/oembed';
   import { BASE_URL as EUROPEANA_DATA_URL } from '@/plugins/europeana/data';
 
   export default {
